separated dna into separate file
